Added Fully Functional BLE Server
Added fully functional BLE server with characteristic for Playback Metadata, Playback Control, Device Time, Message Characteristics and their attributes. Added multithreading and added mutex checks to prevent race conditions. Added implementation for power management based on the time passed as well as if a client is connected to increase the life of the server on a battery operated ESP32.